testing distributed aggregation
